--- main.cpp
@@ -1,21 +1,0 @@
-#include "Server.hpp"
-
-int main(int argc, char *argv[]) {
-  if (argc != 3) {
-    std::cerr << "Usage: " << argv[0] << " <port> <password>" << std::endl;
-    return 1;
-  }
-
-  int port = std::atoi(argv[1]);
-  std::string password = argv[2];
-
-  Server server(port, password);
-  if (!server.start()) {
-    std::cerr << "Failed to start the server." << std::endl;
-    return 1;
-  }
-
-  server.run();
-
-  return 0;
-}
